% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth0DeviceAuth.R
\name{Auth0DeviceAuth-class}
\alias{Auth0DeviceAuth-class}
\alias{Auth0DeviceAuth}
\title{Auth0DeviceAuth}
\description{
Auth0DeviceAuth

Auth0DeviceAuth
}
\details{
Class \code{Auth0DeviceAuth} defines a Auth0DeviceAuth object to interact with Cube Web API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verification_uri}}{character, verification uri}

\item{\code{verification_uri_complete}}{character, verification uri with code}

\item{\code{user_code}}{character, user code}

\item{\code{device_code}}{character, device code}

\item{\code{access_token}}{character access_token}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Auth0DeviceAuth$new()}}
\item \href{#method-get_device_code}{\code{Auth0DeviceAuth$get_device_code()}}
\item \href{#method-get_access_token}{\code{Auth0DeviceAuth$get_access_token()}}
\item \href{#method-finalize}{\code{Auth0DeviceAuth$finalize()}}
\item \href{#method-print}{\code{Auth0DeviceAuth$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
private fields


public fields


Initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth0DeviceAuth$new(
  tenant_url = NULL,
  client_id = NULL,
  scopes = NULL,
  audience = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tenant_url}}{tenant url}

\item{\code{client_id}}{client_id url}

\item{\code{scopes}}{scopes url}

\item{\code{audience}}{audience url}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_code"></a>}}
\if{latex}{\out{\hypertarget{method-get_device_code}{}}}
\subsection{Method \code{get_device_code()}}{
get_device_code get device code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth0DeviceAuth$get_device_code()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response \code{httr::Response}
https://httr.r-lib.org
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_access_token"></a>}}
\if{latex}{\out{\hypertarget{method-get_access_token}{}}}
\subsection{Method \code{get_access_token()}}{
get_access_token
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth0DeviceAuth$get_access_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character access code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth0DeviceAuth$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Auth0DeviceAuth$print()}\if{html}{\out{</div>}}
}

}
}
