% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube.R
\name{Cube-class}
\alias{Cube-class}
\alias{Cube}
\title{The Cube data object}
\description{
The Cube data object

The Cube data object
}
\details{
Class \code{cube} defines a Cube analysis object.
}
\section{Slots}{

\describe{
\item{\code{parameter_file}}{string, full path to YAML file with initialization
parameters}
}}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.database_obj}}{\code{\link{Database}} object}

\item{\code{dataset_id}}{integer dataset id}

\item{\code{accession_ids}}{a character vector of access ids}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.database_obj}}{\code{\link{Database}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cube$new()}}
\item \href{#method-get_metadata}{\code{Cube$get_metadata()}}
\item \href{#method-get_data}{\code{Cube$get_data()}}
\item \href{#method-annotate}{\code{Cube$annotate()}}
\item \href{#method-save_annotated_data}{\code{Cube$save_annotated_data()}}
\item \href{#method-test}{\code{Cube$test()}}
\item \href{#method-finalize}{\code{Cube$finalize()}}
\item \href{#method-print}{\code{Cube$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$new(dataset_id = NULL, accession_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_id}}{integer dataset id}

\item{\code{accession_ids}}{a character vector of access ids}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
get_metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$get_metadata(
  dataset_id = NULL,
  accession_ids = NULL,
  max_number_of_row = 100,
  is_pivot = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_id}}{integer dataset id}

\item{\code{accession_ids}}{a character vector of access ids}

\item{\code{max_number_of_row}}{maximum number of rows returned.
default is 100 and -1 for everything}

\item{\code{is_pivot}}{boolean to indicate if the return data
is pivoted by property_label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
get_data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$get_data(max_number_of_row = 50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_number_of_row}}{maximum number of rows returned.
default is 100 and -1 for everything}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotate"></a>}}
\if{latex}{\out{\hypertarget{method-annotate}{}}}
\subsection{Method \code{annotate()}}{
annotate
annotate derived data objects with a metadata description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$annotate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_annotated_data"></a>}}
\if{latex}{\out{\hypertarget{method-save_annotated_data}{}}}
\subsection{Method \code{save_annotated_data()}}{
save_annotated_data
save the annotated derived data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$save_annotated_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
test method.
This is a test function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$test()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$print()}\if{html}{\out{</div>}}
}

}
}
