% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Disease}
\alias{Cube Metadata Disease}
\alias{disease}
\title{Cube Metadata Disease}
\format{
A Disease metadata with 3 variables:
\describe{
\item{accession_id}{accession id}
\item{hpo_disease_id}{HPO Disease Id }
\item{hpo_disease_term}{HPO Disease Term}
}
}
\usage{
disease
}
\description{
Metadata information for Disease
}
\keyword{datasets}
