% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Individual Mouse}
\alias{Cube Metadata Individual Mouse}
\alias{mouse}
\title{Cube Metadata Individual Mouse}
\format{
A Disease metadata with 15 variables:
\describe{
\item{mouseid}{mouse id}
\item{jumusid}{jumus id}
\item{other}{other}
\item{climbid}{climb id}
\item{dateofbirth}{date of birth}
\item{diet}{diet}
\item{datediet}{date diet}
\item{drugtreatment}{drug treatment}
\item{datedrugstarted}{date drug started}
\item{sex}{sex}
\item{comment}{comment}
\item{partofcohort}{part of cohort}
\item{study}{study}
\item{strain}{strain}
\item{partofstudy}{part of study}
}
}
\usage{
mouse
}
\description{
Metadata information for Individual Mouse
}
\keyword{datasets}
