% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Mouse_Strain}
\alias{Cube Metadata Mouse_Strain}
\alias{mouse_strain}
\title{Cube Metadata Mouse_Strain}
\format{
A Disease metadata with 3 variables:
\describe{
\item{accession_id}{accession id}
\item{strain_name}{strain_name }
\item{classification}{classification}
}
}
\usage{
mouse_strain
}
\description{
Metadata information for Mouse_Strain
}
\keyword{datasets}
