% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CUBE_VERSION}
\alias{CUBE_VERSION}
\title{The constants used in the package}
\format{
An object of class \code{character} of length 1.
}
\usage{
CUBE_VERSION
}
\description{
The constants used in the package
}
\keyword{datasets}
