% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubed.R
\name{cubed}
\alias{cubed}
\title{Cubes a number or a vector/list of numbers}
\usage{
cubed(x)
}
\arguments{
\item{x}{a number or a vector/list to be cubed}
}
\value{
the cube of x in vector format
}
\description{
Cubes a number or a vector/list of numbers
}
\examples{
cubed(3)
cubed(1:3)
}
