% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeR.R
\name{cubeR}
\alias{cubeR}
\title{A CubeR Function}
\usage{
cubeR(dataset = "Assay", accession_ids = "JAXAS00005")
}
\arguments{
\item{dataset}{dataset to access.}

\item{accession_ids}{accession_ids to access.}
}
\description{
This is basic function to retrieve cube data
}
\examples{
cubeR("Assay", "JAXAS00005")
}
\keyword{cubeR}
