% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{Database-class}
\alias{Database-class}
\alias{Database}
\title{Database}
\description{
Database

Database
}
\details{
Class \code{database} defines a database object.
}
\section{Slots}{

\describe{
\item{\code{parameter_file}}{string, full path to YAML file with initialization
parameters}
}}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{host}}{database host}

\item{\code{port}}{database port}

\item{\code{dbname}}{database name}

\item{\code{user}}{database user id}

\item{\code{password}}{user password}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Database$new()}}
\item \href{#method-select}{\code{Database$select()}}
\item \href{#method-test}{\code{Database$test()}}
\item \href{#method-get_column_info}{\code{Database$get_column_info()}}
\item \href{#method-finalize}{\code{Database$finalize()}}
\item \href{#method-print}{\code{Database$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$new(
  host = NULL,
  port = NULL,
  dbname = NULL,
  user = NULL,
  password = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{database host
required}

\item{\code{port}}{database port
required}

\item{\code{dbname}}{database name
required}

\item{\code{user}}{database user id
required}

\item{\code{password}}{database user password
required}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\if{latex}{\out{\hypertarget{method-select}{}}}
\subsection{Method \code{select()}}{
select method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$select(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{a SQL query string
required}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame query_results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
test method.
This is a test function database connection and query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$test()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame query_results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_column_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_column_info}{}}}
\subsection{Method \code{get_column_info()}}{
get_column_info method.
This is a function to get metadata column names
#' @param dataset_id integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$get_column_info(dataset_id)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame of two columns, column name and data type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$print()}\if{html}{\out{</div>}}
}

}
}
