% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_bucket_data}
\alias{get_bucket_data}
\title{get_bucket_data}
\usage{
get_bucket_data(bucket_name, file_name)
}
\arguments{
\item{bucket_name}{character, bucket name}

\item{file_name}{character, file name}
}
\value{
dataframe bucket data
}
\description{
get_bucket_data get bucket data
}
