% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_metadata_query}
\alias{get_metadata_query}
\title{get_metadata_query}
\usage{
get_metadata_query(dataset_id, accession_ids = NULL, max_number_of_row = 100)
}
\arguments{
\item{dataset_id}{integer for the element id in database}

\item{accession_ids}{a vector of character of accession ids to filter on}

\item{max_number_of_row}{maximum number of rows returned.
default is 100 and -1 for everything}
}
\value{
data.frame
}
\description{
This function can return database query string based on the element id
passed in as a parameter.
}
