% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_query_pivot.R
\name{get_metadata_query_pivot}
\alias{get_metadata_query_pivot}
\title{Get the get_metadata_query_pivot}
\usage{
get_metadata_query_pivot(
  dataset_id,
  accession_ids = NULL,
  max_number_of_row = 100,
  column_info
)
}
\arguments{
\item{dataset_id}{integer for the element id in database}

\item{accession_ids}{a vector of character of accession ids to filter on}

\item{max_number_of_row}{maximum number of rows returned.
default is 100 and -1 for everything}

\item{column_info}{data.frame of two columns, column name and data type}
}
\value{
data.frame
}
\description{
This function can return database query string based on the element id
passed in as a parameter.  It pivots the property_label and property value,
grouping by instance id
}
