% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Hello}
\alias{Hello}
\alias{hello}
\title{A Hello World Function}
\usage{
hello(message = "World")
}
\description{
Hello world
}
\details{
This is echo function to test installation.
}
\examples{
hello()
}
\keyword{Hello}
