% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_widget.R
\name{login_widget}
\alias{login_widget}
\title{login widget}
\usage{
login_widget(validation_url, width = 400, height = 100, elementId = NULL)
}
\arguments{
\item{validation_url}{url for validation}
}
\description{
This is login widget for Cube API login.  In RStudio, it will show in viewer
and in R notebook, it shows as cell
}
