% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{response_json_to_data}
\alias{response_json_to_data}
\title{These are the util function shared by the package}
\usage{
response_json_to_data(response)
}
\arguments{
\item{response}{\code{httr::Response} http response}
}
\value{
data \code{jsonlite::fromJSON}. The data could be list or data frame
depend on the json
https://httr.r-lib.org
}
\description{
to_json convert a HTTP json response \code{httr::Response} to R
data \code{jsonlite::fromJSON}.
}
