% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_snp_data.R
\name{show_snp_data}
\alias{show_snp_data}
\title{SNP Grid Data}
\usage{
show_snp_data(num_of_rows = 5)
}
\arguments{
\item{num_of_rows}{number of rows to print.}
}
\description{
This function allows you to access the snp grid data.
}
\examples{
show_snp_data(5)
}
\keyword{cube}
