% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubeAPI.R
\name{CubeAPI-class}
\alias{CubeAPI-class}
\alias{CubeAPI}
\title{CubeAPI}
\description{
CubeAPI

CubeAPI
}
\details{
Class \code{CubeAPI} defines a CubeAPI object to interact with Cube Web API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.token}}{api call token}

\item{\code{cube}}{web app API base}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.token}}{api call token}

\item{\code{cube}}{web app API base}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CubeAPI$new()}}
\item \href{#method-get}{\code{CubeAPI$get()}}
\item \href{#method-get_element_instance}{\code{CubeAPI$get_element_instance()}}
\item \href{#method-get_disclaimer}{\code{CubeAPI$get_disclaimer()}}
\item \href{#method-get_disclaimer_json}{\code{CubeAPI$get_disclaimer_json()}}
\item \href{#method-finalize}{\code{CubeAPI$finalize()}}
\item \href{#method-print}{\code{CubeAPI$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$new(url_base = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url_base}}{cube web app API base}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
get method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$get(endpoint, limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{a get endpoint to call
required}

\item{\code{limit}}{max number of data returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Response \code{httr::Response}
https://httr.r-lib.org
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_element_instance"></a>}}
\if{latex}{\out{\hypertarget{method-get_element_instance}{}}}
\subsection{Method \code{get_element_instance()}}{
get_element_instance method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$get_element_instance(limit = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{max number of data returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Response \code{httr::Response}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_disclaimer"></a>}}
\if{latex}{\out{\hypertarget{method-get_disclaimer}{}}}
\subsection{Method \code{get_disclaimer()}}{
get_disclaimer method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$get_disclaimer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response \code{httr::Response}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_disclaimer_json"></a>}}
\if{latex}{\out{\hypertarget{method-get_disclaimer_json}{}}}
\subsection{Method \code{get_disclaimer_json()}}{
get_disclaimer_json method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$get_disclaimer_json()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json \code{jsonlite::json}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAPI$print()}\if{html}{\out{</div>}}
}

}
}
